+++ view/forum/listTopicsByCategory.php
@@ -13,8 +13,13 @@ if($topics == NULL){
     <p>
         <a href="index.php?ctrl=forum&action=listPostsByTopic&id=<?= $topic->getId() ?>"><?= $topic->getTitle() ?></a> 
         par <?= $topic->getUser() ?>
-        <a href="index.php?ctrl=forum&action=updateTopicForm&id=<?= $topic->getId() ?>"> - Modifier </a>
-        <a href="index.php?ctrl=forum&action=deleteTopic&id=<?= $topic->getId() ?>"> - Supprimer </a>
+
+        <?php if(App\Session::getUser() == $topic->getUser()) { ?>
+
+            <a href="index.php?ctrl=forum&action=updateTopicForm&id=<?= $topic->getId() ?>"> - Modifier </a>
+            <a href="index.php?ctrl=forum&action=deleteTopic&id=<?= $topic->getId() ?>"> - Supprimer </a>
+            
+        <?php } ?>
     </p>
 <?php }
 } ?>
